# target/classes/api.yaml
openapi: 3.0.0
info:
  title: GateKeeper API
  description: Spring Final Project
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local development server
...
paths:
  /residents:
    get:
      summary: To display a List of all residents
      operationId: getResidents
      tags:
        - Admin
      responses:
        '200':
          description: A List of all residents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResidentObject'

    post:
      summary: Create a new resident
      operationId: createResidents
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResident'
      responses:
        '201':
          description: Created a resident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentObject'
        '400':
          description: Bad Request (invalid request payload)
        '500':
          description: Internal Server Error
  /residents/{residentId}:
    parameters:
      - name: residentId
        in: path
        required: true
        schema:
          type: integer
          format: int32

    get:
      summary: Get a resident by ID
      operationId: getResidentsById
      tags:
        - Admin
      responses:
        '200':
          description: The requested resident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentObject'
        '404':
          description: Resident not found

    put:
      summary: Update a resident by ID
      operationId: updateResidents
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResident'
      responses:
        '200':
          description: The updated resident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentObject'
        '400':
          description: Bad Request (invalid request payload)
        '404':
          description: Resident not found
        '500':
          description: Internal Server Error

    delete:
      summary: Delete a resident by ID
      operationId: deleteResidents
      tags:
        - Admin
      responses:
        '204':
          description: No Content (successful deletion)
        '404':
          description: Resident not found
        '500':
          description: Internal Server Error
  /gatekeepers:
    get:
      summary: To display a List of all gatekeepers
      operationId: getGateKeeper
      tags:
        - Admin
      responses:
        '200':
          description: A List of all gatekeepers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GateKeeperObject'

    post:
      summary: Create a new gatekeeper
      operationId: createGateKeeper
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGateKeeper'
      responses:
        '201':
          description: Created a GateKeeper
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateKeeperObject'
        '400':
          description: Bad Request (invalid request payload)
        '500':
          description: Internal Server Error
  /gatekeepers/{gatekeeperId}:
    parameters:
      - name: gatekeeperId
        in: path
        required: true
        schema:
          type: integer
          format: int32

    get:
      summary: Get a gateKeeper by ID
      operationId: getgateKeeperById
      tags:
        - Admin
      responses:
        '200':
          description: The requested Gatekeeper has been retrived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateKeeperObject'
        '404':
          description: GateKeeper not found

    put:
      summary: Update a GateKeeper by ID
      operationId: updateGateKeeper
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGateKeeper'
      responses:
        '200':
          description: The updated GateKeeper
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateKeeperObject'
        '400':
          description: Bad Request (invalid request payload)
        '404':
          description: Resident not found
        '500':
          description: Internal Server Error

    delete:
      summary: Delete a gateKeeper by ID
      operationId: deleteGateKeeper
      tags:
        - Admin
      responses:
        '204':
          description: No Content (successful deletion)
        '404':
          description: Resident not found
        '500':
          description: Internal Server Error
...
  /requests/approve/{requestId}:
    post:
      summary: Approve and save a user to database
      operationId: approveUser
      tags:
        - Admin
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Approved the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserObject'
        '500':
          description: Internal Server Error
  /requests/view:
    get:
      summary: View all the register requests made by users
      operationId: viewRegnRequest
      tags:
        - Admin
      responses:
        '200':
          description: Approved the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserObject'
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    ResidentObject:
      type: object
      properties:
        id:
          type: integer
          format: int32
        residentName:
          type: string
        residenceId:
          type: integer
          format: int32
        aadhar:
          type: integer
          format: int64
        phoneNumber:
          type: integer
          format: int64
    CreateResident:
      type: object
      properties:
        residentName:
          type: string
        residenceId:
          type: integer
          format: int32
        aadhar:
          type: integer
          format: int64
        phoneNumber:
          type: integer
          format: int64
    UpdateResident:
      type: object
      properties:
        residentName:
          type: string
        aadhar:
          type: integer
          format: int64
        phoneNumber:
          type: integer
          format: int64

    #operation for gatekeeper crud
    GateKeeperObject:
      type: object
      properties:
        id:
          type: integer
          format: int32
        gateKeeperName:
          type: string
        gateId:
          type: integer
          format: int32
        aadhar:
          type: integer
          format: int64
        phoneNumber:
          type: integer
          format: int64
    CreateGateKeeper:
      type: object
      properties:
        gateKeeperName:
          type: string
        aadhar:
          type: integer
          format: int64
        phoneNumber:
          type: integer
          format: int64
    UpdateGateKeeper:
      type: object
      properties:
        gateKeeperName:
          type: string
        gateId:
          type: integer
          format: int32
        aadhar:
          type: integer
          format: int64
        phoneNumber:
          type: integer
          format: int64
...
    UserObject:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        userName:
          type: string
        aadhar:
          type: integer
          format: int64
        phoneNumber:
          type: integer
          format: int64
        userType:
          type: string
...